feat: app redirect url application side override
